fix: resolve review comments
