+++ packages/account/src/Sections/Security/ClosingAccount/__tests__/closing-account.spec.tsx
@@ -16,17 +16,17 @@ jest.mock('../closing-account-reason', () => ({
 }));
 
 describe('<ClosingAccountReason />', () => {
-    it('should render the ClosingAccountReason component', async () => {
+    it('should render the ClosingAccountReason component', () => {
         render(<ClosingAccount />);
         expect(screen.getByTestId('dt_closing_account')).toBeInTheDocument();
     });
 
-    it('should render the ClosingAccountSteps component if render_close_account_reason is false', async () => {
+    it('should render the ClosingAccountSteps component if render_close_account_reason is false', () => {
         render(<ClosingAccount />);
         expect(screen.getByText('ClosingAccountSteps')).toBeInTheDocument();
     });
 
-    it('should render the ClosingAccountReason component if render_close_account_reason is true', async () => {
+    it('should render the ClosingAccountReason component if render_close_account_reason is true', () => {
         const spy = jest.spyOn(React, 'useState').mockImplementation(() => [true, jest.fn()]);
         render(<ClosingAccount />);
         expect(screen.getByText('ClosingAccountReason')).toBeInTheDocument();
