+++ packages/account/src/Sections/Security/ClosingAccount/closing-account-pending-conditions/__tests__/closing-account-pending-content.spec.tsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import ClosingAccountPendingContent from '../closing-account-pending-content';
+
+jest.mock('@deriv/components', () => {
+    const original_module = jest.requireActual('@deriv/components');
+    return {
+        ...original_module,
+        Icon: jest.fn(() => <div>mockedIcon</div>),
+    };
+});
+
+describe('ClosingAccountPendingContent', () => {
+    const mock_props: React.ComponentProps<typeof ClosingAccountPendingContent> = {
+        currency_icon: 'USD',
+        value: 100,
+    };
+    it('should render the ClosingAccountPendingContent component', () => {
+        render(<ClosingAccountPendingContent {...mock_props} />);
+        expect(screen.getByText('mockedIcon')).toBeInTheDocument();
+        expect(screen.getByText('100')).toBeInTheDocument();
+    });
+
+    it('should render  title and loginid', () => {
+        const new_props = {
+            ...mock_props,
+            title: 'mock_title',
+            loginid: 'CR123',
+        };
+        render(<ClosingAccountPendingContent {...new_props} />);
+        expect(screen.getByText('mock_title')).toBeInTheDocument();
+        expect(screen.getByText('CR123')).toBeInTheDocument();
+    });
+});
