+++ packages/account/src/Sections/Security/ClosingAccount/closing-account-pending-conditions/closing-account-pending-content.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { Icon, Text } from '@deriv/components';
+
+type TClosingAccountPendingContentProps = {
+    currency_icon: string;
+    loginid?: string;
+    title?: React.ReactNode;
+    value: React.ReactNode;
+};
+
+const ClosingAccountPendingContent = ({ currency_icon, loginid, title, value }: TClosingAccountPendingContentProps) => (
+    <div className='closing-account-error__container'>
+        <div className='closing-account-error__account-details'>
+            <Icon icon={currency_icon} size={24} />
+            <div className='closing-account-error__account'>
+                <Text line_height='s' color='prominent' size='xs'>
+                    {title}
+                </Text>
+                <Text color='prominent' size='xxxs' line_height='xs'>
+                    {loginid}
+                </Text>
+            </div>
+        </div>
+        <Text className='closing-account-error__details' color='prominent' size='xs' line_height='s' align='right'>
+            {value}
+        </Text>
+    </div>
+);
+export default ClosingAccountPendingContent;
