+++ packages/account/src/Sections/Security/ClosingAccount/closing-account-reason.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Redirect } from 'react-router-dom';
-import { FormikValues } from 'formik';
+import { FormikValues, FormikErrors } from 'formik';
 import { Loading, Modal, Text } from '@deriv/components';
 import { routes, WS } from '@deriv/shared';
 import { localize, Localize } from '@deriv/translations';
-import ClosingAccountHasPendingConditions from './closing-account-has-pending-conditions';
+import { TClosingAccountFormValues } from 'Types';
+import ClosingAccountHasPendingConditions from './closing-account-pending-conditions/closing-account-has-pending-conditions';
 import ClosingAccountReasonForm from './closing-account-reason-form';
 import ClosingAccountWarningModal from './closing-account-warning-modal';
 import ClosingAccountGeneralErrorContent from './closing-account-general-error-content';
@@ -13,16 +14,11 @@ type TClosingAccountReasonProps = {
     onBackClick: () => void;
 };
 
-type Error = {
-    characters_limits?: string;
-    empty_reason?: string;
-};
+const CHARACTER_LIMIT = 110;
+const MAX_ALLOWED_REASONS = 3;
 
-const character_limit_no = 110;
-const max_allowed_reasons = 3;
-
-const preparingReason = (values: FormikValues) => {
-    let selected_reasons = selectedReasons(values)
+const formatReasonsForCloseAccount = (values: TClosingAccountFormValues) => {
+    let selected_reasons = selectedReasonsForCloseAccount(values)
         .map(val => val[0])
         .toString();
     const is_other_trading_platform__has_value = !!values.other_trading_platforms.length;
@@ -37,39 +33,37 @@ const preparingReason = (values: FormikValues) => {
     return selected_reasons.replace(/(\r\n|\n|\r)/gm, ' ');
 };
 
-const selectedReasons = (values: FormikValues) => {
-    return Object.entries(values).filter(
+const selectedReasonsForCloseAccount = (values: TClosingAccountFormValues) =>
+    Object.entries(values).filter(
         ([key, value]) => !['other_trading_platforms', 'do_to_improve'].includes(key) && value
     );
-};
 
 const ClosingAccountReason = ({ onBackClick }: TClosingAccountReasonProps) => {
     const [is_account_closed, setIsAccountClosed] = React.useState(false);
     const [is_loading, setIsLoading] = React.useState(false);
     const [is_modal_open, setIsModalOpen] = React.useState(false);
-    const [which_modal_should_render, setWhichModalShouldRender] = React.useState<string>('');
-    const [reason, setReason] = React.useState<string>('');
+    const [modal_type, setModalType] = React.useState<string>('');
+    const [reason, setReason] = React.useState('');
     const [is_checkbox_disabled, setIsCheckboxDisabled] = React.useState(false);
     const [total_checkbox_checked, setTotalCheckboxChecked] = React.useState(0);
-    const [remaining_characters, setRemainingCharacters] = React.useState(character_limit_no);
+    const [remaining_characters, setRemainingCharacters] = React.useState(CHARACTER_LIMIT);
     const [total_accumulated_characters, setTotalAccumulatedCharacters] = React.useState(0);
     const [api_error_message, setApiErrorMessage] = React.useState('');
     const [details, setDetails] = React.useState();
 
     React.useEffect(() => {
-        if (total_checkbox_checked === max_allowed_reasons) setIsCheckboxDisabled(true);
+        if (total_checkbox_checked === MAX_ALLOWED_REASONS) setIsCheckboxDisabled(true);
         else if (is_checkbox_disabled) setIsCheckboxDisabled(false);
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [total_checkbox_checked]);
+    }, [total_checkbox_checked, is_checkbox_disabled]);
 
-    const validateFields = (values: FormikValues) => {
-        const error: Error = {};
-        const selected_reason_count = selectedReasons(values).length;
+    const validateFields = (values: TClosingAccountFormValues) => {
+        const error: FormikErrors<FormikValues> = {};
+        const selected_reason_count = selectedReasonsForCloseAccount(values).length;
         const text_inputs_length = (values.other_trading_platforms + values.do_to_improve).length;
-        let remaining_chars = character_limit_no - text_inputs_length;
+        let remaining_chars = CHARACTER_LIMIT - text_inputs_length;
 
         if (selected_reason_count) {
-            const final_value = preparingReason(values);
+            const final_value = formatReasonsForCloseAccount(values);
             remaining_chars = remaining_chars >= 0 ? remaining_chars : 0;
 
             if (!/^[a-zA-Z0-9.,'\-\s]*$/.test(final_value)) {
@@ -85,11 +79,11 @@ const ClosingAccountReason = ({ onBackClick }: TClosingAccountReasonProps) => {
         return error;
     };
 
-    const handleSubmitForm = (values: FormikValues) => {
-        const final_reason = preparingReason(values);
+    const handleSubmitForm = (values: TClosingAccountFormValues) => {
+        const final_reason = formatReasonsForCloseAccount(values);
 
         setIsModalOpen(true);
-        setWhichModalShouldRender('warning_modal');
+        setModalType('warning_modal');
         setReason(final_reason);
     };
 
@@ -115,7 +109,7 @@ const ClosingAccountReason = ({ onBackClick }: TClosingAccountReasonProps) => {
         const value = e.target.value;
         const is_delete_action = old_value.length > value.length;
 
-        if ((remaining_characters <= 0 || total_accumulated_characters >= character_limit_no) && !is_delete_action) {
+        if ((remaining_characters <= 0 || total_accumulated_characters >= CHARACTER_LIMIT) && !is_delete_action) {
             e.preventDefault();
         } else {
             onChange(e);
@@ -131,10 +125,10 @@ const ClosingAccountReason = ({ onBackClick }: TClosingAccountReasonProps) => {
     };
 
     const getModalTitle = () => {
-        if (which_modal_should_render === 'error_modal') return <Localize i18n_default_text='An error occurred' />;
-        else if (which_modal_should_render === 'inaccessible_modal')
+        if (modal_type === 'error_modal') return <Localize i18n_default_text='An error occurred' />;
+        else if (modal_type === 'inaccessible_modal')
             return <Localize i18n_default_text='Inaccessible MT5 account(s)' />;
-        else if (which_modal_should_render !== 'warning_modal') return <Localize i18n_default_text='Action required' />;
+        else if (modal_type !== 'warning_modal') return <Localize i18n_default_text='Action required' />;
         return '';
     };
 
@@ -160,7 +154,7 @@ const ClosingAccountReason = ({ onBackClick }: TClosingAccountReasonProps) => {
                 return 'error_modal';
             };
 
-            setWhichModalShouldRender(getModalToRender());
+            setModalType(getModalToRender());
             setDetails(errorDetails);
             setApiErrorMessage(message);
             setIsModalOpen(true);
@@ -176,8 +170,8 @@ const ClosingAccountReason = ({ onBackClick }: TClosingAccountReasonProps) => {
         <div className='closing-account-reasons'>
             <Text weight='bold' size='xs' className='closing-account-reasons__title' as='p'>
                 <Localize
-                    i18n_default_text='Please tell us why you’re leaving. (Select up to {{ allowed_reasons }} reasons.'
-                    values={{ allowed_reasons: max_allowed_reasons }}
+                    i18n_default_text='Please tell us why you’re leaving. (Select up to {{ allowed_reasons }} reasons.)'
+                    values={{ allowed_reasons: MAX_ALLOWED_REASONS }}
                 />
             </Text>
             <ClosingAccountReasonForm
@@ -185,7 +179,7 @@ const ClosingAccountReason = ({ onBackClick }: TClosingAccountReasonProps) => {
                 onSubmit={handleSubmitForm}
                 is_checkbox_disabled={is_checkbox_disabled}
                 onChangeCheckbox={handleChangeCheckbox}
-                character_limit_no={character_limit_no}
+                character_limit_no={CHARACTER_LIMIT}
                 onInputChange={handleInputChange}
                 onInputPaste={handleInputPaste}
                 remaining_characters={remaining_characters}
@@ -197,16 +191,16 @@ const ClosingAccountReason = ({ onBackClick }: TClosingAccountReasonProps) => {
                 toggleModal={() => setIsModalOpen(!is_modal_open)}
                 title={getModalTitle()}
             >
-                {which_modal_should_render === 'warning_modal' && (
+                {modal_type === 'warning_modal' && (
                     <ClosingAccountWarningModal
                         closeModal={() => setIsModalOpen(false)}
                         startDeactivating={startDeactivating}
                     />
                 )}
-                {which_modal_should_render === 'AccountHasPendingConditions' && (
+                {modal_type === 'AccountHasPendingConditions' && (
                     <ClosingAccountHasPendingConditions details={details} onBackClick={onBackClick} />
                 )}
-                {which_modal_should_render === 'inaccessible_modal' && (
+                {modal_type === 'inaccessible_modal' && (
                     <ClosingAccountGeneralErrorContent
                         message={api_error_message}
                         onClick={() => setIsModalOpen(false)}
