+++ packages/account/src/Sections/Security/ClosingAccount/closing-account.tsx
@@ -3,7 +3,7 @@ import ClosingAccountSteps from './closing-account-steps';
 import ClosingAccountReason from './closing-account-reason';
 
 const ClosingAccount = () => {
-    const [render_close_account_reason, setRenderCloseAccountReason] = React.useState<boolean>(false);
+    const [render_close_account_reason, setRenderCloseAccountReason] = React.useState(false);
     const redirectToReasons = () => {
         setRenderCloseAccountReason(true);
     };
@@ -14,9 +14,9 @@ const ClosingAccount = () => {
     return (
         <div className='closing-account' data-testid='dt_closing_account'>
             {render_close_account_reason ? (
-                <ClosingAccountReason onBackClick={() => redirectToSteps()} />
+                <ClosingAccountReason onBackClick={redirectToSteps} />
             ) : (
-                <ClosingAccountSteps redirectToReasons={() => redirectToReasons()} />
+                <ClosingAccountSteps redirectToReasons={redirectToReasons} />
             )}
         </div>
     );
